Added spellmap filtering and sorting
Filter the spellmap items to be limited to just the selected
spellbook items and sort them by level and then by school.
This will enable linking to be more easily impelemented in
the near future. Also fixed a bug that would turn off the
checkbox in the list if the label of the spellbook checkbox
was clicked.